HARP-7373: Move text style caching to new TextStyleCache in TextElementsRenderer.
- Consolidate text style caches into a TextStyleCache instance owned by TextElementsRenderer.
- Remove duplicated code in PoiManager and TileGeometryCreator to get text styles.
- Simplify TextElementsRenderer initialization. The object it's always available in MapView,
only its initialization is delayed until the view is configured to render labels and there's
labels available from some data source.
